--- packages/plasma-hope/src/components/Button/Button.mixins.ts
@@ -1,3 +1,0 @@
-import { getButtonSizesMixin, buttonSizes, buttonTypography } from '@salutejs/plasma-core';
-
-export const applySizes = getButtonSizesMixin(buttonSizes, buttonTypography);
