+++ packages/plasma-hope/src/components/Button/index.tsx
@@ -1,5 +1,5 @@
 export { Button } from './Button';
-export type { ButtonProps } from './Button';
+export type { ButtonProps } from './types';
 
 export { buttonViews } from './Button.props';
 export type { ButtonView } from './Button.props';
